+++ packages/capacitor/src/builders/add/schema.d.ts
@@ -1,5 +1,0 @@
-import { JsonObject } from '@angular-devkit/core';
-
-export interface AddBuilderSchema extends JsonObject {
-  platform: string;
-}