+++ packages/capacitor/src/builders/command/schema.d.ts
@@ -1,0 +1,6 @@
+import { JsonObject } from '@angular-devkit/core';
+
+export interface CommandBuilderSchema extends JsonObject {
+  command: 'add' | 'copy' | 'open' | 'sync' | 'update';
+  platform: string;
+}