+++ packages/capacitor/src/builders/sync/schema.d.ts
@@ -1,5 +1,0 @@
-import { JsonObject } from '@angular-devkit/core';
-
-export interface SyncBuilderSchema extends JsonObject {
-  platform: string;
-}